Regenerated javadoc
Link: Adjusted to use notion of frame
	Also made certain methods final (had to adjust WrapperLink)

Scheduler: Can get the decision structure of a node (see
getDecisionStructure(...)).
	DecisionStructure: pickRandomInput(...) and getReservedCapacity(...),
to facilitate StatisticalScheduler. Currently just a stub, as the
reservation is not being done anywhere.
	
Constants: Used for placing constants. Includes FRAME_SIZE = 1000.

Flow: Container for a sequence of hops, as well as capacity needed for
the flow.

Network: To be used to create host nodes that generate flow and
switches. This will be maintaining the flow reservations within each
decision structure (unimplemented, will do it soon when I have thought
about how to go about those reservations).